Move terminate_cluster back into cluster module
Fixes #27
