+++ cumulus/starcluster/tasks/cluster.py
@@ -69,5 +69,37 @@ def start_cluster(cluster, base_url=None, log_write_url=None, girder_token=None,
         if config_filepath and os.path.exists(config_filepath):
             os.remove(config_filepath)
 
+@app.task
+@cumulus.starcluster.logging.capture
+def terminate_cluster(cluster, base_url=None, log_write_url=None, girder_token=None):
+    name = cluster['name']
+    cluster_id = cluster['_id']
+    config_id = cluster['configId']
+    log_write_url = '%s/clusters/%s/log' % (base_url, cluster_id)
+    config_url = '%s/starcluster-configs/%s' % (base_url, config_id)
+    status_url = '%s/clusters/%s' % (base_url, cluster_id)
+
+
+    config_filepath = None
+    try:
+        config_filepath = _write_config_file(girder_token, config_url)
+        config = starcluster.config.StarClusterConfig(config_filepath)
+        config.load()
+        cm = config.get_cluster_manager()
+
+
+        with logstdout():
+            cm.terminate_cluster(name, force=True)
+
+        # Now update the status of the cluster
+        headers = {'Girder-Token':  girder_token}
+        r = requests.patch(status_url, headers=headers, json={'status': 'terminated'})
+        _check_status(r)
+    except starcluster.exception.ClusterDoesNotExist:
+        r = requests.patch(status_url, headers=headers, json={'status': 'terminated'})
+        _check_status(r)
+    finally:
+        if config_filepath and os.path.exists(config_filepath):
+            os.remove(config_filepath)
